ボスの弾発射から、スタン→再移動まで
