+++ Project/Game/Objects/GameCameraState/GameCameraState.cpp
@@ -7,6 +7,9 @@
 #include "Objects/Player/Player.h"
 #include "GraphicManager/Light/DirectionalLight.h"
 
+//function
+#include"Easing/EasingFunction.h"
+#include"FrameManager/Time.h"
 
 void GameCameraState::Initialize() {
 
@@ -16,23 +19,68 @@ void GameCameraState::Initialize() {
 	pTransform_->rotateOrder = QUATERNION;
 	pTransform_->Update();
 
+	easeTMax_ = 0.7f;
+	zoomOutMax_ = -70.0f;
+	zoomInMax_ = -50.0f;
 }
 
 void GameCameraState::Update() {
-
-
-	pLight_->SetDirection(-pGameCamera_->GetPosition().Normalize());
+	//プレイヤーのゲージMaxでカメラズームアウト
+	if (pPlayer_->GetisPowerUp()) {
+		SetBehaviorZoomOut();
+	}
+	//ズームイン
+	else if (!pPlayer_->GetisPowerUp()) {
+		SetBehaviorZoomIn();
+	}
+	//振る舞い管理
+	BehaviorManagement();
 }
 
 void GameCameraState::LastUpdate() {
 
 }
 
+//振る舞い関数
+void GameCameraState::RootInit() {
+
+}
+void GameCameraState::RootUpdate() {
+
+}
+
+void  GameCameraState::ZoomInInit() {
+	//ズームアウトのカメラの位置になるよう初期化
+	pTransform_->position.z = zoomOutMax_;
+	easeT_ = 0.0f;
+}
+void  GameCameraState::ZoomInUpdate() {
+	easeT_ += Time::DeltaTime();//デルタタイムにしたい
+	if (easeT_ >= easeTMax_) {
+		easeT_ = easeTMax_;
+	}
+	//カメラを引く
+	pTransform_->position.z = EaseOutBack<float>(zoomOutMax_, zoomInMax_, easeT_, easeTMax_);
+}
+
+void  GameCameraState::ZoomOutInit() {
+	//通常状態のカメラの位置になるよう初期化
+	pTransform_->position.z = zoomInMax_;
+	easeT_ = 0.0f;
+}
+
+void  GameCameraState::ZoomOutUpdate() {
+	easeT_ += Time::DeltaTime();//デルタタイムにしたい
+	if (easeT_ >= easeTMax_) {
+		easeT_ = easeTMax_;
+	}
+	//カメラを引く
+	pTransform_->position.z = EaseOutBack<float>(zoomInMax_, zoomOutMax_, easeT_, easeTMax_);
+}
 
 
 void GameCameraState::Debug() {
-
-
+	
 	if(ImGui::TreeNodeEx("player var", ImGuiTreeNodeFlags_DefaultOpen)) {
 
 
@@ -40,11 +88,66 @@ void GameCameraState::Debug() {
 	}
 
 	if(ImGui::TreeNodeEx("camera var", ImGuiTreeNodeFlags_DefaultOpen)) {
-
+		ImGui::DragFloat("zoomOutMax", &zoomOutMax_);
+		ImGui::DragFloat("zoomInMax", &zoomInMax_);
 
 		ImGui::TreePop();
 	}
 
+}
+
+
+//振る舞い管理
+void  GameCameraState::BehaviorManagement() {
+	if (behaviorRequest_) {
+		// 振る舞いを変更する
+		behavior_ = behaviorRequest_.value();
+		// 各振る舞いごとの初期化を実行
+		switch (behavior_) {
+		case Behavior::kRoot:
+		default:
+			RootInit();
+			break;
+		case Behavior::kZoomIn:
+			ZoomInInit();
+			break;
+		case Behavior::kZoomOut:
+			ZoomOutInit();
+			break;
+		}
+		// 振る舞いリクエストをリセット
+		behaviorRequest_ = std::nullopt;
+	}
+	// 振る舞い更新
+	switch (behavior_) {
+	case Behavior::kRoot:
+	default:
+		RootUpdate();
+		break;
+	case Behavior::kZoomIn:
+		ZoomInUpdate();
+		break;
+	case Behavior::kZoomOut:
+		ZoomOutUpdate();
+		break;
+
+	}
+
+}
+
+//ズームアウトにセット
+void GameCameraState::SetBehaviorZoomOut() {
+	if (behavior_ != Behavior::kZoomOut) {
+		behaviorRequest_ = Behavior::kZoomOut;
+	}
+}
+
+//ズームインにセット
+void GameCameraState::SetBehaviorZoomIn() {
+	//ズームアウトの時にInに戻す
+	if (behavior_ == Behavior::kZoomOut) {
+		behaviorRequest_ = Behavior::kZoomIn;
+	}
 }
