+++ Project/ONEngine/GameObjectManager/BaseGameObject.cpp
@@ -56,8 +56,8 @@ const Vec3 BaseGameObject::GetPosition() const {
 /// 親のセット
 /// ===================================================
 void BaseGameObject::SetParent(Transform* parent) {
-	pTranform_->SetParent(parent);
-	pTranform_->AddChild(GetTransform()); //- 相手の子供に自身を追加
+	pTransform_->SetParent(parent);
+	parent->AddChild(GetTransform()); //- 相手の子供に自身を追加
 }
 
 
@@ -73,12 +73,12 @@ void BaseGameObject::ParentCancel(bool isLocalToWorld) {
 	if(!parent) { return; }
 
 	if(isLocalToWorld) {
-		pTranform_->Update();
-		pTranform_->position = GetPosition();
+		pTransform_->Update();
+		pTransform_->position = GetPosition();
 	}
 
-	pTranform_->parent_ = nullptr;	
-
+	pTransform_->parent_ = nullptr;
+	parent->SubChild(pTransform_);
 }
